backend/requirements.txt
# requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Redis==0.4.0
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0